+++ familyhelper-ui-component/familyhelper-ui-component-api/src/api/settingrepo/settingNode.js
@@ -1,6 +1,4 @@
-import {
-    del, get, patch, post,
-} from '@/util/http';
+import {get, post} from '@/util/http';
 
 export function exists(key) {
     return get('settingrepo', `setting-node/${key}/exists/`, {});
@@ -8,30 +6,6 @@ export function exists(key) {
 
 export function inspect(key) {
     return get('settingrepo', `setting-node/${key}/`, {});
-}
-
-export function insert(key, value, remark) {
-    return post('settingrepo', 'setting-node/', {
-        key: {
-            string_id: key,
-        },
-        value,
-        remark,
-    });
-}
-
-export function remove(key) {
-    return del('settingrepo', `setting-node/${key}/`, {});
-}
-
-export function update(key, value, remark) {
-    return patch('settingrepo', 'setting-node/', {
-        key: {
-            string_id: key,
-        },
-        value,
-        remark,
-    });
 }
 
 export function all(page, rows) {
@@ -56,11 +30,11 @@ export function operateInspect(category, args) {
     });
 }
 
-export function operatePut(category, args, value, remark) {
-    return post('settingrepo', 'setting-node/put/', {
+export function operateInit(category, args, type, remark) {
+    return post('settingrepo', 'setting-node/init/', {
         category,
         args,
-        value,
+        type,
         remark
     });
 }
@@ -71,4 +45,3 @@ export function operateRemove(category, args) {
         args
     });
 }
-