+++ familyhelper-ui-view/familyhelper-ui-view-pc/src/views/items/life/activityTemplateManagement/ActivityTemplateEditPanel.vue
@@ -179,8 +179,7 @@ from '@/views/items/life/activityTemplateManagement/ActivityTemplateDriverPanel.
 import FloatyDialog from '@/components/layout/FloatyDialog.vue';
 import FileEditPanel from '@/views/items/miscellaneous/fileEditor/FileEditPanel.vue';
 
-import { operateInspect, operatePut } from '@/api/settingrepo/settingNode';
-import { currentTimestamp, formatTimestamp } from '@/util/timestamp';
+import { operateInspect, operatePut } from '@/api/settingrepo/textNode';
 import resolveResponse from '@/util/response';
 
 import { LIFE_ACTIVITY_TEMPLATE_FILE } from '@/views/items/miscellaneous/fileEditor/fileTypeConstants';
@@ -377,10 +376,11 @@ export default {
         return;
       }
       resolveResponse(
-        operatePut(this.upsc,
+        operatePut(
+          this.upsc,
           [this.me],
           JSON.stringify(this.getUserPreference()),
-          `更新时间: ${formatTimestamp(currentTimestamp())}`),
+        ),
       ).then(() => {
         this.$message({
           showClose: true,