+++ familyhelper-ui-view/familyhelper-ui-view-pc/src/views/items/note/noteManagement/ItemEditPanel.vue
@@ -108,8 +108,7 @@ import ItemAttachmentPanel from '@/views/items/note/noteManagement/ItemAttachmen
 import FloatyDialog from '@/components/layout/FloatyDialog.vue';
 import FileEditPanel from '@/views/items/miscellaneous/fileEditor/FileEditPanel.vue';
 
-import { operateInspect, operatePut } from '@/api/settingrepo/settingNode';
-import { currentTimestamp, formatTimestamp } from '@/util/timestamp';
+import { operateInspect, operatePut } from '@/api/settingrepo/textNode';
 import resolveResponse from '@/util/response';
 
 import { NOTE_ATTACHMENT_FILE } from '@/views/items/miscellaneous/fileEditor/fileTypeConstants';
@@ -310,10 +309,11 @@ export default {
         return;
       }
       resolveResponse(
-        operatePut(this.upsc,
+        operatePut(
+          this.upsc,
           [this.me],
           JSON.stringify(this.getUserPreference()),
-          `更新时间: ${formatTimestamp(currentTimestamp())}`),
+        ),
       ).then(() => {
         this.$message({
           showClose: true,